requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.2
scikit-learn>=0.24.0
matplotlib>=3.3.3
seaborn>=0.11.0
tqdm>=4.54.1
Pillow>=8.0.1
pytorch-grad-cam>=1.3.9
thop>=0.0.31-2005241907 